Small tweaks to ssr
