+++ packages/dom-expressions/src/server/ssr.js
@@ -96,18 +96,10 @@ export function escape(s, attr) {
   return left < s.length ? out + s.substring(left) : out;
 }
 
-export function generateHydrationScript({
-  eventNames = ["click", "input", "blur"],
-  streaming,
-  resolved
-} = {}) {
-  let s = `(()=>{_$HYDRATION={events:[],completed:new WeakSet};const t=e=>e&&e.hasAttribute&&(e.hasAttribute("data-hk")&&e||t(e.host&&e.host instanceof Node?e.host:e.parentNode)),e=e=>{let o=e.composedPath&&e.composedPath()[0]||e.target,s=t(o);s&&!_$HYDRATION.completed.has(s)&&_$HYDRATION.events.push([s,e])};["${eventNames.join(
-    '","'
-  )}"].forEach(t=>document.addEventListener(t,e))})();`;
-  if (streaming) {
-    s += `(()=>{const e=_$HYDRATION,r={};let o=0;e.resolveResource=((e,o)=>{const t=r[e];if(!t)return r[e]=o;delete r[e],t(o)}),e.loadResource=(()=>{const e=++o,t=r[e];if(!t){let o,t=new Promise(e=>o=e);return r[e]=o,t}return delete r[e],Promise.resolve(t)})})();`;
-  }
-  if (resolved)
-    s += `_$HYDRATION.resources = JSON.parse('${JSON.stringify(_$HYDRATION.resources || {})}');`;
-  return s;
+export function resolveSSRNode(node) {
+  if (Array.isArray(node)) return node.map(resolveSSRNode).join("");
+  const t = typeof node;
+  if (node && t === "object") return resolveSSRNode(node.t);
+  if (t === "function") return resolveSSRNode(node());
+  return t === "string" ? node : JSON.stringify(node);
 }
